refactor: example driver class to common library
Both Bigtable and Storage had nearly identical code to drive the
examples, this was done intentionally because at the time the common
library was in the `g-c-cpp-common` repository. This PR just cleans that
up.
