+++ google/cloud/bigtable/examples/bigtable_examples_common_test.cc
@@ -30,138 +30,6 @@ using ::testing::Invoke;
 using ::testing::ReturnRef;
 using ::testing::StartsWith;
 
-TEST(BigtableExamplesCommon, Simple) {
-  int test_calls = 0;
-  Example example({
-      {"test",
-       [&](std::vector<std::string> const& args) {
-         ++test_calls;
-         if (args.empty()) throw Usage("test-usage");
-         ASSERT_EQ(2, args.size());
-         EXPECT_EQ("a0", args[0]);
-         EXPECT_EQ("a1", args[1]);
-       }},
-  });
-  char const* argv[] = {"argv0", "test", "a0", "a1"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 0);
-  EXPECT_EQ(2, test_calls);
-}
-
-TEST(BigtableExamplesCommon, AutoRunDisabled) {
-  google::cloud::testing_util::ScopedEnvironment env(
-      "GOOGLE_CLOUD_CPP_AUTO_RUN_EXAMPLES", "no");
-  int test_calls = 0;
-  Example example({
-      {"test", [&](std::vector<std::string> const&) { ++test_calls; }},
-  });
-  char const* argv[] = {"argv0"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(1, test_calls);
-}
-
-TEST(BigtableExamplesCommon, AutoRunMissing) {
-  google::cloud::testing_util::ScopedEnvironment env(
-      "GOOGLE_CLOUD_CPP_AUTO_RUN_EXAMPLES", "yes");
-  int test_calls = 0;
-  Example example({
-      {"test", [&](std::vector<std::string> const&) { ++test_calls; }},
-  });
-  char const* argv[] = {"argv0"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(1, test_calls);
-}
-
-TEST(BigtableExamplesCommon, AutoRun) {
-  google::cloud::testing_util::ScopedEnvironment env(
-      "GOOGLE_CLOUD_CPP_AUTO_RUN_EXAMPLES", "yes");
-  int test_calls = 0;
-  int auto_calls = 0;
-  Example example({
-      {"test", [&](std::vector<std::string> const&) { ++test_calls; }},
-      {"auto", [&](std::vector<std::string> const&) { ++auto_calls; }},
-  });
-  char const* argv[] = {"argv0"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 0);
-  EXPECT_EQ(1, test_calls);
-  EXPECT_EQ(1, auto_calls);
-}
-
-TEST(BigtableExamplesCommon, CommandNotFound) {
-  int test_calls = 0;
-  Example example({
-      {"test", [&](std::vector<std::string> const&) { ++test_calls; }},
-  });
-  char const* argv[] = {"argv0", "wrong-name"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(1, test_calls);
-}
-
-TEST(BigtableExamplesCommon, CommandUsage) {
-  int test_calls = 0;
-  Example example({
-      {"test",
-       [&](std::vector<std::string> const& args) {
-         ++test_calls;
-         if (args.empty()) throw Usage("test-usage");
-       }},
-  });
-  char const* argv[] = {"argv0", "test"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(2, test_calls);
-}
-
-TEST(BigtableExamplesCommon, CommandError) {
-  int test_calls = 0;
-  Example example({
-      {"test",
-       [&](std::vector<std::string> const& args) {
-         ++test_calls;
-         if (args.empty()) throw Usage("test-usage");
-         throw std::runtime_error("some problem");
-       }},
-  });
-  char const* argv[] = {"argv0", "test", "a0"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(2, test_calls);
-}
-
-TEST(BigtableExamplesCommon, CheckEnvironmentVariablesNormal) {
-  google::cloud::testing_util::ScopedEnvironment test_a("TEST_A", "a");
-  google::cloud::testing_util::ScopedEnvironment test_b("TEST_B", "b");
-  EXPECT_NO_THROW(CheckEnvironmentVariablesAreSet({"TEST_A", "TEST_B"}));
-}
-
-TEST(BigtableExamplesCommon, CheckEnvironmentVariablesNotSet) {
-  google::cloud::testing_util::ScopedEnvironment test_a("TEST_A", {});
-  EXPECT_THROW(
-      try {
-        CheckEnvironmentVariablesAreSet({"TEST_A"});
-      } catch (std::runtime_error const& ex) {
-        EXPECT_THAT(ex.what(), HasSubstr("TEST_A"));
-        throw;
-      },
-      std::runtime_error);
-}
-
-TEST(BigtableExamplesCommon, CheckEnvironmentVariablesSetEmpty) {
-  google::cloud::testing_util::ScopedEnvironment test_a("TEST_A", "");
-  EXPECT_THROW(
-      try {
-        CheckEnvironmentVariablesAreSet({"TEST_A"});
-      } catch (std::runtime_error const& ex) {
-        EXPECT_THAT(ex.what(), HasSubstr("TEST_A"));
-        throw;
-      },
-      std::runtime_error);
-}
-
 TEST(BigtableExamplesCommon, RunAdminIntegrationTestsEmulator) {
   google::cloud::testing_util::ScopedEnvironment emulator(
       "BIGTABLE_EMULATOR_HOST", "localhost:9090");
