+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -26,6 +26,7 @@ google_cloud_cpp_common_unit_tests = [
     "internal/big_endian_test.cc",
     "internal/compiler_info_test.cc",
     "internal/env_test.cc",
+    "internal/example_driver_test.cc",
     "internal/filesystem_test.cc",
     "internal/format_time_point_test.cc",
     "internal/future_impl_test.cc",
