+++ google/cloud/storage/examples/storage_examples_common.cc
@@ -21,85 +21,10 @@ namespace cloud {
 namespace storage {
 namespace examples {
 
-Example::Example(std::map<std::string, CommandType> commands)
-    : commands_(std::move(commands)) {
-  // Force each command to generate its Usage string, so we can provide a good
-  // usage string for the whole program.
-  for (auto const& kv : commands_) {
-    if (kv.first == "auto") continue;
-    try {
-      kv.second({"--help"});
-    } catch (Usage const& u) {
-      full_usage_ += "    ";
-      full_usage_ += u.what();
-      full_usage_ += "\n";
-    }
-  }
-}
-
-int Example::Run(int argc, char const* const argv[]) try {
-  bool auto_run =
-      google::cloud::internal::GetEnv("GOOGLE_CLOUD_CPP_AUTO_RUN_EXAMPLES")
-          .value_or("") == "yes";
-  if (argc == 1 && auto_run) {
-    auto entry = commands_.find("auto");
-    if (entry == commands_.end()) {
-      PrintUsage(argv[0], "Requested auto run but there is no 'auto' command");
-      return 1;
-    }
-    entry->second({});
-    return 0;
-  }
-
-  if (argc < 2) {
-    PrintUsage(argv[0], "Missing command");
-    return 1;
-  }
-
-  std::string const command_name = argv[1];
-  auto command = commands_.find(command_name);
-  if (commands_.end() == command) {
-    PrintUsage(argv[0], "Unknown command: " + command_name);
-    return 1;
-  }
-
-  command->second({argv + 2, argv + argc});
-
-  return 0;
-} catch (Usage const& u) {
-  PrintUsage(argv[0], u.what());
-  return 1;
-} catch (std::exception const& ex) {
-  std::cerr << "Standard C++ exception raised: " << ex.what() << "\n";
-  return 1;
-}
-
-void Example::PrintUsage(std::string const& cmd, std::string const& msg) {
-  auto last_slash = cmd.find_last_of('/');
-  auto program = cmd.substr(last_slash + 1);
-  std::cerr << msg << "\nUsage: " << program << " <command> [arguments]\n\n"
-            << "Commands:\n"
-            << full_usage_ << "\n";
-}
-
 bool UsingTestbench() {
   return !google::cloud::internal::GetEnv("CLOUD_STORAGE_TESTBENCH_ENDPOINT")
               .value_or("")
               .empty();
-}
-
-void CheckEnvironmentVariablesAreSet(std::vector<std::string> const& vars) {
-  for (auto const& var : vars) {
-    auto const value = google::cloud::internal::GetEnv(var.c_str());
-    if (!value) {
-      throw std::runtime_error("The " + var +
-                               " environment variable is not set");
-    }
-    if (value->empty()) {
-      throw std::runtime_error("The " + var +
-                               " environment variable has an empty value");
-    }
-  }
 }
 
 std::string MakeRandomBucketName(google::cloud::internal::DefaultPRNG& gen,
