+++ google/cloud/storage/examples/storage_examples_common_test.cc
@@ -24,152 +24,6 @@ namespace examples {
 using ::testing::HasSubstr;
 using ::testing::StartsWith;
 
-TEST(StorageExamplesCommon, Simple) {
-  int test_calls = 0;
-  Example example({
-      {"test",
-       [&](std::vector<std::string> const& args) {
-         ++test_calls;
-         if (args.empty()) throw Usage("test-usage");
-         if (args[0] == "--help") throw Usage("usage with --help");
-         ASSERT_EQ(2, args.size());
-         EXPECT_EQ("a0", args[0]);
-         EXPECT_EQ("a1", args[1]);
-       }},
-  });
-  char const* argv[] = {"argv0", "test", "a0", "a1"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 0);
-  EXPECT_EQ(2, test_calls);
-}
-
-TEST(StorageExamplesCommon, AutoRunDisabled) {
-  google::cloud::testing_util::ScopedEnvironment env(
-      "GOOGLE_CLOUD_CPP_AUTO_RUN_EXAMPLES", "no");
-  int test_calls = 0;
-  Example example({
-      {"test", [&](std::vector<std::string> const&) { ++test_calls; }},
-  });
-  char const* argv[] = {"argv0"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(1, test_calls);
-}
-
-TEST(StorageExamplesCommon, AutoRunMissing) {
-  google::cloud::testing_util::ScopedEnvironment env(
-      "GOOGLE_CLOUD_CPP_AUTO_RUN_EXAMPLES", "yes");
-  int test_calls = 0;
-  Example example({
-      {"test", [&](std::vector<std::string> const&) { ++test_calls; }},
-  });
-  char const* argv[] = {"argv0"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(1, test_calls);
-}
-
-TEST(StorageExamplesCommon, AutoRun) {
-  google::cloud::testing_util::ScopedEnvironment env(
-      "GOOGLE_CLOUD_CPP_AUTO_RUN_EXAMPLES", "yes");
-  int test_calls = 0;
-  int auto_calls = 0;
-  Example example({
-      {"test", [&](std::vector<std::string> const&) { ++test_calls; }},
-      {"auto", [&](std::vector<std::string> const&) { ++auto_calls; }},
-  });
-  char const* argv[] = {"argv0"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 0);
-  EXPECT_EQ(1, test_calls);
-  EXPECT_EQ(1, auto_calls);
-}
-
-TEST(StorageExamplesCommon, CommandNotFound) {
-  int test_calls = 0;
-  Example example({
-      {"test", [&](std::vector<std::string> const&) { ++test_calls; }},
-  });
-  char const* argv[] = {"argv0", "wrong-name"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(1, test_calls);
-}
-
-TEST(StorageExamplesCommon, CommandUsage) {
-  int test_calls = 0;
-  Example example({
-      {"test",
-       [&](std::vector<std::string> const& args) {
-         ++test_calls;
-         if (args.empty()) throw Usage("test-usage");
-       }},
-  });
-  char const* argv[] = {"argv0", "test"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(2, test_calls);
-}
-
-TEST(StorageExamplesCommon, CommandError) {
-  int test_calls = 0;
-  Example example({
-      {"test",
-       [&](std::vector<std::string> const& args) {
-         ++test_calls;
-         if (args.empty()) throw Usage("test-usage");
-         if (args[0] == "--help") throw Usage("usage with --help");
-         throw std::runtime_error("some problem");
-       }},
-  });
-  char const* argv[] = {"argv0", "test", "a0"};
-  int argc = sizeof(argv) / sizeof(argv[0]);
-  EXPECT_EQ(example.Run(argc, argv), 1);
-  EXPECT_EQ(2, test_calls);
-}
-
-TEST(StorageExamplesCommon, UsingTestbenchTrue) {
-  google::cloud::testing_util::ScopedEnvironment env(
-      "CLOUD_STORAGE_TESTBENCH_ENDPOINT", "http://localhost:9090");
-  EXPECT_TRUE(UsingTestbench());
-}
-
-TEST(StorageExamplesCommon, UsingTestbenchFalse) {
-  google::cloud::testing_util::ScopedEnvironment env(
-      "CLOUD_STORAGE_TESTBENCH_ENDPOINT", {});
-  EXPECT_FALSE(UsingTestbench());
-}
-
-TEST(StorageExamplesCommon, CheckEnvironmentVariablesNormal) {
-  google::cloud::testing_util::ScopedEnvironment test_a("TEST_A", "a");
-  google::cloud::testing_util::ScopedEnvironment test_b("TEST_B", "b");
-  EXPECT_NO_THROW(CheckEnvironmentVariablesAreSet({"TEST_A", "TEST_B"}));
-}
-
-TEST(StorageExamplesCommon, CheckEnvironmentVariablesNotSet) {
-  google::cloud::testing_util::ScopedEnvironment test_a("TEST_A", {});
-  EXPECT_THROW(
-      try {
-        CheckEnvironmentVariablesAreSet({"TEST_A"});
-      } catch (std::runtime_error const& ex) {
-        EXPECT_THAT(ex.what(), HasSubstr("TEST_A"));
-        throw;
-      },
-      std::runtime_error);
-}
-
-TEST(StorageExamplesCommon, CheckEnvironmentVariablesSetEmpty) {
-  google::cloud::testing_util::ScopedEnvironment test_a("TEST_A", "");
-  EXPECT_THROW(
-      try {
-        CheckEnvironmentVariablesAreSet({"TEST_A"});
-      } catch (std::runtime_error const& ex) {
-        EXPECT_THAT(ex.what(), HasSubstr("TEST_A"));
-        throw;
-      },
-      std::runtime_error);
-}
-
 TEST(StorageExamplesCommon, RandomBucket) {
   auto generator = google::cloud::internal::DefaultPRNG(std::random_device{}());
   auto const actual_1 = MakeRandomBucketName(generator, "test-prefix-");
